refactor: move contents in `ipc` to `manager`
To hide functions.
